Fixed warnings in NSDateFormatter and UIColor+Zest

+++ desiccant/Classes/Zest/UIKit/UIColor+Zest.m
@@ -419,7 +419,7 @@ static NSLock *crayolaNameCacheLock;
 }
 
 - (NSString *)hexStringFromColor {
-	return [NSString stringWithFormat:@"%0.6X", self.rgbHex];
+	return [NSString stringWithFormat:@"%0.6X", (unsigned int)self.rgbHex];
 }
 
 - (NSString *)closestColorNameFor: (const char *) aColorDatabase {
